+++ modules/bluetooth.nix
@@ -1,0 +1,7 @@
+_: {
+  hardware.bluetooth = {
+    enable = true;
+    powerOnBoot = false;
+  };
+  services.blueman.enable = true;
+}