+++ modules/gnome.nix
@@ -2,7 +2,8 @@
   lib,
   pkgs,
   ...
-}: {
+}:
+{
   config = {
     # Enable the X11 windowing system.
     services.xserver.enable = true;
@@ -11,7 +12,7 @@
     ];
 
     # Enable the GNOME Desktop Environment.
-    services.displayManager.gdm.enable = true;
+    # services.displayManager.gdm.enable = true;
     services.desktopManager.gnome.enable = true;
 
     # Used to generate thumbnails.
@@ -102,6 +103,7 @@
     xdg.mime.enable = true;
     # Find the desktop file in nix store with: find /nix/store/ -name "*application_name*desktop"
     xdg.mime.defaultApplications = {
+      "inode/directory" = "nautilus.desktop";
       "audio/aac" = "vlc.desktop";
       "audio/mpeg" = "vlc.desktop";
       "audio/ogg" = "vlc.desktop";